Changed the style.css file
in order for the confirm button to be green I had to change the order of
the CSS rules. I had to move the confirm button styles to the end of the
file to ensure they override any previous styles.
In order to prevent overriding of the the child paragraph's styles I
deleted thetype selector and moved the child class selector to the end.